refactor(hardcover): Insert records with batches

+++ booklore-api/src/main/java/org/booklore/model/dto/BookIdentifier.java
@@ -2,7 +2,7 @@ package org.booklore.model.dto;
 
 public interface BookIdentifier {
     Integer getProgressId();
-    Long getBookId();
+    int getBookId();
     String getIsbn10();
     String getIsbn13();
     String getHardcoverBookId();
